pyproject.toml
[tool.poetry]
name = "Crypto_Wallet"
version = "0.1.0"
description = "CryptoWallet project"
authors = ["Dima Rubets <rubetsdima100@gmail.com>"]

[tool.poetry.dependencies]
# Compatible Python versions
python = ">=3.8,<4.0"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
wand = "^0.6.11"
sqlalchemy-imageattach = "^1.1.0"
sqlalchemy-utils = "^0.41.1"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
pydantic = {extras = ["email"], version = "^2.1.1"}
emails = "^0.6"
jinja2 = "^3.1.2"
python-environ = "^0.4.54"
python-multipart = "^0.0.6"
dependency-injector = "^4.41.0"
asyncpg = "^0.28.0"
pytest = "^7.4.0"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"
aio-pika = "^9.2.1"
propan = {extras = ["async-rabbit"], version = "^0.1.5.17"}
web3 = "^6.8.0"
python-socketio = "^5.8.0"

[tool.pytest.ini_options]

asyncio_mode="auto"
